+++ internal/web/static/app.js
@@ -18,7 +18,7 @@ import "./filehist.js";
 import "./review.js";
 import { applyStoredWidths } from "./resize.js";
 import { applyGraphMode, loadCommits, renderCommits } from "./commits.js";
-import { refreshNoteCounts } from "./files.js";
+import { applyDiffView, refreshNoteCounts } from "./files.js";
 import { focusPane } from "./keys.js";
 import { fetchHealth } from "./bigrepo.js";
 import "./settings.js";
@@ -56,6 +56,7 @@ async function applyStoredLayout() {
   applyStoredWidths(ui.sidebar_width, ui.files_width);
   if (ui.sidebar_hidden) applySidebarHidden(true);
   applyGraphMode(ui.graph);
+  applyDiffView(ui.diff_view);
 }